+++ easy_post/lib/easy_post_v2.dart
@@ -1,0 +1,5 @@
+export './src/post.dart';
+export './src/post.service.dart';
+export './src/screens/post.edit.screen.dart';
+export './src/widgets/post.list_tile.dart';
+export './src/widgets/post.list_view.dart';