+++ easy_post/lib/src/defines.dart
@@ -1,0 +1,3 @@
+import 'package:firebase_auth/firebase_auth.dart';
+
+User? get currentUser => FirebaseAuth.instance.currentUser;