+++ easy_post/lib/src/post.service.dart
@@ -1,0 +1,43 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:easy_post_v2/easy_post_v2.dart';
+import 'package:easy_post_v2/src/screens/post.list.screen.dart';
+import 'package:flutter/material.dart';
+
+class PostService {
+  static PostService? _instance;
+
+  static PostService get instance => _instance ??= PostService._();
+  PostService._();
+
+  bool initialized = false;
+  CollectionReference get col => FirebaseFirestore.instance.collection('posts');
+
+  init() {
+    initialized = true;
+  }
+
+  Future<DocumentReference?> showPostEditScreen({
+    required BuildContext context,
+    required String? category,
+    Post? post,
+  }) {
+    return showGeneralDialog<DocumentReference?>(
+      context: context,
+      pageBuilder: (_, __, ___) {
+        return PostEditScreen(category: category);
+      },
+    );
+  }
+
+  Future showPostListScreen({
+    required BuildContext context,
+    Post? post,
+  }) {
+    return showGeneralDialog(
+      context: context,
+      pageBuilder: (_, __, ___) {
+        return const PostListScreen();
+      },
+    );
+  }
+}