+++ easy_post/lib/src/widgets/post.list_tile.dart
@@ -1,0 +1,16 @@
+import 'package:easy_post_v2/easy_post_v2.dart';
+import 'package:flutter/material.dart';
+
+class PostListTile extends StatelessWidget {
+  const PostListTile({super.key, required this.post});
+  final Post post;
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: Text(post.title),
+      subtitle: Text(post.content),
+      // onTap: ()=> ,
+    );
+  }
+}