+++ easy_task/lib/easy_task.dart
@@ -35,6 +35,7 @@ export 'src/functions.dart';
 // Unit Test
 export 'unit_tests/task/task.unit_test.dart';
 
+
 // FOR REVIEW ===========
 // NOTE: User should not be exported
 //       Because this should be handled