+++ example/lib/main.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:easy_locale/easy_locale.dart';
-import 'package:easychat/easychat.dart';
+import 'package:easy_post_v2/easy_post_v2.dart';
 import 'package:easyuser/easyuser.dart';
 import 'package:example/etc/zone_error_handler.dart';
 // import 'package:example/firebase_options.dart';
@@ -40,9 +40,13 @@ class MyAppState extends State<MyApp> {
   void initState() {
     super.initState();
     UserService.instance.init();
+    PostService.instance.init();
 
     SchedulerBinding.instance.addPostFrameCallback((timeStamp) {
-      ChatService.instance.showChatRoomEditScreen(globalContext);
+      // ChatService.instance.showChatRoomEditScreen(globalContext);
+      // PostService.instance.showPostEditScreen(context: globalContext);
+
+      PostService.instance.showPostListScreen(context: globalContext);
     });
   }