+++ easy_storage/CHANGELOG.md
@@ -1,5 +1,9 @@
+## 0.0.5
+* Add. widgets for uploading.
+
+
 ## 0.0.4
-* Add. UploadImage widget.
+* Add. ImageUpload widget.
 
 ## 0.0.3
 * Flutter Firebase SDK version adjustment.